Added logic for update ticket and added hook
