+++ src/app/_components/UI/DragAndDrop/DragDropLoader/dragdroploadr.scss
@@ -1,0 +1,11 @@
+.dragdroparealoader{
+    display: flex;
+}
+
+.taskclumnwrapperloader{
+    margin: 8px;
+    display: flex;
+    width: 100%;
+    border-radius: 10px;
+    min-height: 60vh!important;
+}
