+++ src/app/dashboard/manageticket/[id]/ticketmanage.scss
@@ -1,17 +1,8 @@
-.dragdroparealoader{
-    display: flex;
-}
+
 .projecttitle{
     color: #0D062D;
 font-family: Inter;
 font-size: 18px;
 font-style: normal;
 font-weight: 600;
-}
-.taskclumnwrapperloader{
-    margin: 8px;
-    display: flex;
-    width: 100%;
-    border-radius: 10px;
-    min-height: 60vh!important;
 }
