+++ src/app/globals.css
@@ -113,13 +113,15 @@ body{
 [data-tooltip]:after {
   z-index: 1000;
   padding: 8px;
-  width: 160px;
+  min-width: 140px;
+  max-width: 300px;
   background-color: #000;
   background-color: hsla(0, 0%, 20%, 0.9);
   color: #fff;
   content: attr(data-tooltip);
   font-size: 14px;
   line-height: 1.2;
+  white-space: nowrap;
 }
 
 /* Directions */
